+++ tzatziki-spring-jpa/src/test/java/com/decathlon/tzatziki/steps/TestApplicationSteps.java
@@ -9,8 +9,8 @@ import org.springframework.context.ApplicationContextInitializer;
 import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.test.context.ContextConfiguration;
 import org.testcontainers.containers.PostgreSQLContainer;
+import org.testcontainers.containers.wait.strategy.Wait;
 
-import java.net.Socket;
 import java.util.List;
 import java.util.Map;
 
@@ -26,30 +26,18 @@ public class TestApplicationSteps {
     }
 
     private static final PostgreSQLContainer<?> postgres =
-            new PostgreSQLContainer<>("postgres:16").withTmpFs(Map.of("/var/lib/postgresql/data", "rw"));
+            new PostgreSQLContainer<>("postgres:16").withTmpFs(Map.of("/var/lib/postgresql/data", "rw")).waitingFor(Wait.forListeningPort());
 
     static class Initializer implements ApplicationContextInitializer<ConfigurableApplicationContext> {
 
         public void initialize(ConfigurableApplicationContext configurableApplicationContext) {
             postgres.start();
-            waitForPort(postgres.getHost(), postgres.getFirstMappedPort());
             TestPropertyValues.of(
                     "spring.datasource.url=" + postgres.getJdbcUrl(),
                     "spring.datasource.username=" + postgres.getUsername(),
                     "spring.datasource.password=" + postgres.getPassword()
             ).applyTo(configurableApplicationContext.getEnvironment());
             SpringJPASteps.schemasToClean = List.of("public", "library", "store");
-        }
-
-        private static void waitForPort(String host, int port) {
-            for (int i = 0; i < 30; i++) {
-                try (Socket s = new Socket(host, port)) {
-                    return;
-                } catch (Exception e) {
-                    try { Thread.sleep(500); } catch (InterruptedException ie) { Thread.currentThread().interrupt(); }
-                }
-            }
-            throw new IllegalStateException("Port " + host + ":" + port + " not reachable after 15s");
         }
     }
 }
